refactor(bench): split generators into focused modules
